Avanzamento svilippo app

+++ app/Http/Controllers/CertificateController.php
@@ -16,6 +16,8 @@ class CertificateController extends Controller
      */
     public function index(Athlete $athlete)
     {
+        $this->authorize('xxx');
+
         if (request()->ajax()) {
             return datatables()->eloquent($athlete->certificates())
             ->addColumn('action', function ($certificate) use($athlete){
@@ -32,6 +34,9 @@ class CertificateController extends Controller
      */
     public function create(Athlete $athlete)
     {
+
+        $this->authorize('xxx');
+
         $certificate = new Certificate();
         $certificate->is_current = 1;
         return view('backend.athletes.certificates.create', compact('athlete', 'certificate'));
@@ -42,6 +47,9 @@ class CertificateController extends Controller
      */
     public function store(AthleteCertificatesRequest $request, Athlete $athlete)
     {
+
+        $this->authorize('xxx');
+
         $athlete->certificate()->create($request->validated());
         Utility::flashMessage();
         
@@ -53,7 +61,7 @@ class CertificateController extends Controller
      */
     public function show(Certificate $certificate)
     {
-        //
+        $this->authorize('xxx');
     }
 
     /**
@@ -61,6 +69,8 @@ class CertificateController extends Controller
      */
     public function edit(Athlete $athlete, Certificate $certificate)
     {
+        $this->authorize('xxx');
+
         return view('backend.athletes.certificates.edit', compact('athlete', 'certificate'));
     }
 
@@ -69,6 +79,8 @@ class CertificateController extends Controller
      */
     public function update(AthleteCertificatesRequest $request, Athlete $athlete, Certificate $certificate)
     {
+        $this->authorize('xxx');
+
         $certificate->update($request->validated());
         Utility::flashMessage();
         return redirect(route('athletes.certificates.index', $athlete));
@@ -79,6 +91,9 @@ class CertificateController extends Controller
      */
     public function destroy(Athlete $athlete, Certificate $certificate)
     {
+        $this->authorize('xxx');
+        
+
         $certificate->delete();
         Utility::flashMessage();
         return redirect(route('athletes.certificates.index', $athlete));
